shrink-vm: 增加一个判断，在无 snap 的虚拟机上忽略 shrinking snap 的命令

+++ shrink-vm.sh
@@ -18,8 +18,10 @@ fi
 
 # Removes old revisions of snaps
 # CLOSE ALL SNAPS BEFORE RUNNING THIS
-set -eu
-snap list --all | awk '/disabled/{print $1, $3}' |
-    while read snapname revision; do
-        sudo snap remove "$snapname" --revision="$revision"
-    done
+if [ -d /snap ]; then
+    set -eu
+    snap list --all | awk '/disabled/{print $1, $3}' |
+        while read snapname revision; do
+            sudo snap remove "$snapname" --revision="$revision"
+        done
+fi
